5.4 The raise statement and custom exceptions
I have used exceptions since the beginning of learning
the Python language, and what it is was explained only now.
and did not explain plainly,
but described what the concept of 'exception' is.
how does it work? what does it serve? what can be done about it?
can not understand anything.
I may have to take another course to learn OOP.